# docs/customizations/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "Mayan EDMS API",
    "description": "Free Open Source Electronic Document Management System",
    "license": {
      "name": "Apache 2.0"
    },
    "version": "v4"
  },
  "host": "localhost:8080",
  "schemes": [
    "http"
  ],
  "basePath": "/api/v4",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "Basic": {
      "type": "basic"
    }
  },
  "security": [
    {
      "Basic": []
    }
  ],
  "paths": {
...
    "/roles/": {
      "get": {
        "operationId": "roles_list",
        "description": "Returns a list of all the roles.",
        "parameters": [
          {
            "name": "_ordering",
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      },
      "post": {
        "operationId": "roles_create",
        "description": "Create a new role.",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      },
      "parameters": []
    },
    "/roles/{role_id}/": {
      "get": {
        "operationId": "roles_read",
        "description": "Return the details of the selected role.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      },
      "put": {
        "operationId": "roles_update",
        "description": "Edit the selected role.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      },
      "patch": {
        "operationId": "roles_partial_update",
        "description": "Edit the selected role.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      },
      "delete": {
        "operationId": "roles_delete",
        "description": "Delete the selected role.",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      },
      "parameters": [
        {
          "name": "role_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
    "/roles/{role_id}/permissions/": {
      "get": {
        "operationId": "roles_permissions_list",
        "description": "Returns a list of all the permissions granted to a particular role.",
        "parameters": [
          {
            "name": "_ordering",
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      },
      "parameters": [
        {
          "name": "role_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
  },
...
}